+++ src/test/java/StatsServiceTests.java
@@ -4,7 +4,7 @@ import ru.netology.stats.StatsService;
 
 public class StatsServiceTests {
 
-    int[] salesInMonth = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+    long[] salesInMonth = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
 
     @Test
